Rename src/polls/consts.ts -> src/consts/polls.ts
