+++ test/stores/room-list/previews/PollStartEventPreview-test.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import { MatrixEvent } from "matrix-js-sdk";
 
-import { IPollAnswer } from "../../../../src/polls/consts";
+import { IPollAnswer } from "../../../../src/consts/polls";
 import { PollStartEventPreview } from "../../../../src/stores/room-list/previews/PollStartEventPreview";
 import { MatrixClientPeg } from "../../../../src/MatrixClientPeg";
